requirements.txt
aiofiles~=23.1.0
pyserial-asyncio==0.6
